requirements.txt
numpy~=1.22.4
pandas~=1.4.3
matplotlib~=3.5.2
torch~=1.13.1
tqdm~=4.64.0
PyYAML~=6.0
librosa~=0.9.2
fast-bss-eval~=0.1.4
torchaudio~=0.13.1
HyperPyYAML~=1.0.1
torchmetrics~=0.7.0
musdb~=0.4.0
pytorch-lightning~=1.9.5
asteroid~=0.6.0
speechbrain~=0.5.14
demucs~=4.0.0
omegaconf~=2.3.0
SoundFile~=0.10.3.post1
pyloudnorm~=0.1.0
scipy~=1.9.0
wandb~=0.13.1
museval~=0.4.1
julius~=0.2.7
openunmix~=1.2.1
einops~=0.4.1
hydra_colorlog~=1.2.0
hydra-core~=1.3.2